requirements.txt
# Core LLM Providers
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0
groq>=0.4.0

# Database & ORM
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Background Processing
apscheduler>=3.10.0
flask>=3.0.0
flask-cors>=4.0.0

# Web Scraping & Research
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.0.0
html5lib>=1.1

# Knowledge Base & Vector DB
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Document Processing
python-docx>=1.1.0
PyPDF2>=3.0.0
reportlab>=4.0.0
markdown>=3.5.0

# Email & Communication
sendgrid>=6.10.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0
rich>=13.7.0
typer>=0.9.0
diff-match-patch>=20230430

# AWS S3 (optional, for storage service)
boto3>=1.28.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
locust>=2.17.0
faker>=19.0.0
httpx>=0.24.0

# Web Framework
streamlit>=1.28.0
streamlit-aggrid>=0.3.4
streamlit-option-menu>=0.3.6
plotly>=5.17.0

# API Enhancements
flask-limiter>=3.5.0
flask-jwt-extended>=4.6.0
flasgger>=0.9.7
PyJWT>=2.8.0

# Monitoring & Analytics
prometheus-client>=0.19.0
sentry-sdk>=1.40.0
python-json-logger>=2.0.7
psutil>=5.9.0

# Performance & Caching
redis>=5.0.0
cachetools>=5.3.0
sqlalchemy-utils>=0.41.0
memory-profiler>=0.61.0

# Security & Compliance
cryptography>=41.0.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
bleach>=6.0.0
safety>=2.3.0
bandit>=1.7.0

# Legacy (may be used by existing code)
langchain>=0.1.0
langchain-openai>=0.0.2
pydantic>=2.0.0

